# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cytellect private API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/revisions/{rid}/workspace-selection": {
      "get": {
        "summary": "Saved Selection",
        "operationId": "saved_selection_v1_revisions__rid__workspace_selection_get",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSelection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{wid}/region-cohorts": {
      "post": {
        "summary": "Assemble",
        "operationId": "assemble_v1_workspaces__wid__region_cohorts_post",
        "parameters": [
          {
            "name": "wid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionCohortRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CohortSource": {
        "properties": {
          "field_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Field Id"
          },
          "revision_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Revision Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field_id",
          "revision_id"
        ],
        "title": "CohortSource"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegionCohortRequest": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/CohortSource"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Sources"
          },
          "metadata": {
            "patternProperties": {
              "^[A-Za-z0-9_-]+$": {
                "$ref": "#/components/schemas/RegionFieldMetadata"
              }
            },
            "propertyNames": {
              "maxLength": 80,
              "minLength": 1
            },
            "type": "object",
            "maxProperties": 100,
            "minProperties": 1,
            "title": "Metadata"
          },
          "expected_active_revision_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Active Revision Id"
          },
          "workspace_selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceSelection"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sources",
          "metadata",
          "expected_active_revision_id"
        ],
        "title": "RegionCohortRequest"
      },
...
      "RegionFieldMetadata": {
        "properties": {
          "condition": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          },
          "experimental_unit": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Experimental Unit"
          },
          "sample": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample"
          },
          "acquisition_date": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquisition Date"
          },
          "pair": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Pair"
          },
          "repeat_length": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Repeat Length"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RegionFieldMetadata"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkspaceSelection": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Version"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceSelectionEntry"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Entries"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "entries"
        ],
        "title": "WorkspaceSelection"
      },
      "WorkspaceSelectionEntry": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Id"
          },
          "field_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Id"
          },
          "revision_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Id"
          },
          "exclusion_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclusion Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id"
        ],
        "title": "WorkspaceSelectionEntry"
      },
...
    }
  }
}